+++ src/main/Core/BrowserWindow/BackgroundMaterial/index.ts
