+++ src/main/Core/BrowserWindow/Contract/BackgroundMaterialProvider.ts
@@ -1,0 +1,5 @@
+import type { BrowserWindowConstructorOptions } from "electron";
+
+export interface BrowserWindowBackgroundMaterialProvider {
+    get(): BrowserWindowConstructorOptions["backgroundMaterial"];
+}
