+++ src/main/Core/BrowserWindow/Contract/VibrancyProvider.ts
@@ -1,0 +1,5 @@
+import type { Vibrancy } from "./Vibrancy";
+
+export interface BrowserWindowVibrancyProvider {
+    get(): Vibrancy | null;
+}
