+++ src/main/Core/BrowserWindow/Contract/index.ts
@@ -1,0 +1,3 @@
+export * from "./BackgroundMaterialProvider";
+export * from "./Vibrancy";
+export * from "./VibrancyProvider";
