+++ src/main/Core/BrowserWindow/Vibrancy/VibrancyProvider.test.ts
