+++ src/main/Core/BrowserWindow/Vibrancy/VibrancyProvider.ts
@@ -1,22 +1,7 @@
 import type { SettingsManager } from "@Core/SettingsManager";
+import type { Vibrancy, BrowserWindowVibrancyProvider as VibrancyProviderInterface } from "../Contract";
 
-type Vibrancy =
-    | "content"
-    | "fullscreen-ui"
-    | "header"
-    | "hud"
-    | "menu"
-    | "popover"
-    | "selection"
-    | "sheet"
-    | "sidebar"
-    | "titlebar"
-    | "tooltip"
-    | "under-page"
-    | "under-window"
-    | "window";
-
-export class VibrancyProvider {
+export class VibrancyProvider implements VibrancyProviderInterface {
     public constructor(private readonly settingsManager: SettingsManager) {}
 
     public get(): Vibrancy | null {
