+++ src/main/Core/BrowserWindow/Vibrancy/index.ts
