+++ src/main/Core/BrowserWindow/index.ts
@@ -1,1 +1,2 @@
 export * from "./BrowserWindowModule";
+export * from "./Contract";
