+++ src/main/Core/SearchWindow/ActionHandler/NavigateToActionHandler.test.ts
