+++ src/main/Core/SearchWindow/ActionHandler/index.ts
