+++ src/main/Core/SearchWindow/AppIconFilePathResolver/AppIconFilePathResolver.test.ts
