+++ src/main/Core/SearchWindow/AppIconFilePathResolver/AppIconFilePathResolver.ts
