+++ src/main/Core/SearchWindow/AppIconFilePathResolver/index.ts
