+++ src/main/Core/SearchWindow/BrowserWindowConstructorOptionsProvider/BrowserWindowConstructorOptionsProvider.ts
