+++ src/main/Core/SearchWindow/BrowserWindowConstructorOptionsProvider/DefaultBrowserWindowConstructorOptionsProvider.test.ts
