+++ src/main/Core/SearchWindow/BrowserWindowConstructorOptionsProvider/LinuxBrowserWindowConstructorOptionsProvider.test.ts
