+++ src/main/Core/SearchWindow/BrowserWindowConstructorOptionsProvider/LinuxBrowserWindowConstructorOptionsProvider.ts
