+++ src/main/Core/SearchWindow/BrowserWindowConstructorOptionsProvider/MacOsBrowserWindowConstructorOptionsProvider.ts
@@ -1,19 +1,19 @@
+import type { BrowserWindowVibrancyProvider } from "@Core/BrowserWindow";
 import type { BrowserWindowConstructorOptions } from "electron";
-import type { BackgroundMaterialProvider } from "./BackgroundMaterial";
 import type { BrowserWindowConstructorOptionsProvider } from "./BrowserWindowConstructorOptionsProvider";
 
-export class WindowsBrowserWindowConstructorOptionsProvider implements BrowserWindowConstructorOptionsProvider {
+export class MacOsBrowserWindowConstructorOptionsProvider implements BrowserWindowConstructorOptionsProvider {
     public constructor(
         private readonly defaultOptions: BrowserWindowConstructorOptions,
-        private readonly backgroundMaterialProvider: BackgroundMaterialProvider,
+        private readonly vibrancyProvider: BrowserWindowVibrancyProvider,
     ) {}
 
     public get(): Electron.BrowserWindowConstructorOptions {
         return {
             ...this.defaultOptions,
             ...{
-                autoHideMenuBar: true,
-                backgroundMaterial: this.backgroundMaterialProvider.get(),
+                vibrancy: this.vibrancyProvider.get() ?? undefined,
+                backgroundColor: "rgba(0,0,0,0)",
             },
         };
     }
