+++ src/main/Core/SearchWindow/BrowserWindowConstructorOptionsProvider/WindowsBrowserWindowConstructorOptionsProvider.ts
@@ -1,19 +1,19 @@
+import type { BrowserWindowBackgroundMaterialProvider } from "@Core/BrowserWindow";
 import type { BrowserWindowConstructorOptions } from "electron";
 import type { BrowserWindowConstructorOptionsProvider } from "./BrowserWindowConstructorOptionsProvider";
-import type { VibrancyProvider } from "./Vibrancy";
 
-export class MacOsBrowserWindowConstructorOptionsProvider implements BrowserWindowConstructorOptionsProvider {
+export class WindowsBrowserWindowConstructorOptionsProvider implements BrowserWindowConstructorOptionsProvider {
     public constructor(
         private readonly defaultOptions: BrowserWindowConstructorOptions,
-        private readonly vibrancyProvider: VibrancyProvider,
+        private readonly backgroundMaterialProvider: BrowserWindowBackgroundMaterialProvider,
     ) {}
 
     public get(): Electron.BrowserWindowConstructorOptions {
         return {
             ...this.defaultOptions,
             ...{
-                vibrancy: this.vibrancyProvider.get() ?? undefined,
-                backgroundColor: "rgba(0,0,0,0)",
+                autoHideMenuBar: true,
+                backgroundMaterial: this.backgroundMaterialProvider.get(),
             },
         };
     }
