+++ src/main/Core/SearchWindow/BrowserWindowConstructorOptionsProvider/defaultWindowSize.ts
