+++ src/main/Core/SearchWindow/BrowserWindowConstructorOptionsProvider/index.ts
@@ -1,8 +1,8 @@
-export * from "./BackgroundMaterial";
+export * from "../../BrowserWindow/BackgroundMaterial";
+export * from "../../BrowserWindow/Vibrancy";
 export * from "./BrowserWindowConstructorOptionsProvider";
 export * from "./DefaultBrowserWindowConstructorOptionsProvider";
 export * from "./defaultWindowSize";
 export * from "./LinuxBrowserWindowConstructorOptionsProvider";
 export * from "./MacOsBrowserWindowConstructorOptionsProvider";
-export * from "./Vibrancy";
 export * from "./WindowsBrowserWindowConstructorOptionsProvider";
