+++ src/main/Core/SearchWindow/BrowserWindowToggler/BrowserWindowToggler.test.ts
