+++ src/main/Core/SearchWindow/BrowserWindowToggler/BrowserWindowToggler.ts
