+++ src/main/Core/SearchWindow/BrowserWindowToggler/index.ts
