+++ src/main/Core/SearchWindow/index.ts
@@ -1,0 +1,1 @@
+export * from "./SearchWindowModule";
