+++ src/main/index.ts
@@ -70,6 +70,7 @@ if (!app.requestSingleInstanceLock()) {
     Core.SettingsReaderModule.bootstrap(dependencyRegistry);
     Core.SettingsWriterModule.bootstrap(dependencyRegistry);
     Core.SettingsManagerModule.bootstrap(dependencyRegistry);
+    Core.BrowserWindowModule.bootstrap(dependencyRegistry);
     Core.ShellModule.bootstrap(dependencyRegistry);
     Core.DockModule.bootstrap(dependencyRegistry);
     await Core.ImageGeneratorModule.bootstrap(dependencyRegistry);
@@ -91,7 +92,7 @@ if (!app.requestSingleInstanceLock()) {
     await Core.ExtensionManagerModule.bootstrap(dependencyRegistry);
 
     // Windows
-    await Core.BrowserWindowModule.bootstrap(dependencyRegistry);
+    await Core.SearchWindowModule.bootstrap(dependencyRegistry);
     await Core.SettingsWindowModule.bootstrap(dependencyRegistry);
 
     Core.RescanOrchestratorModule.bootstrap(dependencyRegistry);
