+++ src/renderer/Core/Settings/Pages/index.tsx
@@ -32,27 +32,27 @@ export const settingsPages: SettingsPage[] = [
     {
         translation: { key: "title", namespace: "settingsGeneral" },
         relativePath: "general",
-        absolutePath: "/settings/general",
+        absolutePath: "/general",
         element: <General />,
         icon: <SettingsRegular />,
     },
     {
         translation: { key: "title", namespace: "settingsWindow" },
         relativePath: "window",
-        absolutePath: "/settings/window",
+        absolutePath: "/window",
         element: <Window />,
         icon: <WindowRegular />,
     },
     {
         translation: { key: "title", namespace: "settingsAppearance" },
         relativePath: "appearance",
-        absolutePath: "/settings/appearance",
+        absolutePath: "/appearance",
         element: <Appearance />,
         icon: <PaintBrushRegular />,
     },
     {
         translation: { key: "title", namespace: "settingsKeyboardAndMouse" },
-        absolutePath: "/settings/keyboard-and-mouse",
+        absolutePath: "/keyboard-and-mouse",
         element: <KeyboardAndMouse />,
         relativePath: "keyboard-and-mouse",
         icon: <KeyboardMouse16Regular />,
@@ -60,35 +60,35 @@ export const settingsPages: SettingsPage[] = [
     {
         translation: { key: "title", namespace: "settingsSearchEngine" },
         relativePath: "search-engine",
-        absolutePath: "/settings/search-engine",
+        absolutePath: "/search-engine",
         element: <SearchEngine />,
         icon: <SearchRegular />,
     },
     {
         translation: { key: "title", namespace: "settingsFavorites" },
         relativePath: "favorites",
-        absolutePath: "/settings/favorites",
+        absolutePath: "/favorites",
         element: <Favorites />,
         icon: <StarRegular />,
     },
     {
         translation: { key: "title", namespace: "settingsExtensions" },
         relativePath: "extensions",
-        absolutePath: "/settings/extensions",
+        absolutePath: "/extensions",
         element: <Extensions />,
         icon: <AppsAddInRegular />,
     },
     {
         translation: { key: "title", namespace: "settingsAbout" },
         relativePath: "about",
-        absolutePath: "/settings/about",
+        absolutePath: "/about",
         element: <About />,
         icon: <InfoRegular />,
     },
     {
         translation: { key: "title", namespace: "settingsDebug" },
         relativePath: "debug",
-        absolutePath: "/settings/debug",
+        absolutePath: "/debug",
         element: <Debug />,
         icon: <BugRegular />,
     },
